+++ src/utils.ts
@@ -452,12 +452,11 @@ export const StatusMap = {
 export type HTTPStatusName = keyof typeof StatusMap
 
 export const signCookie = async (val: string, secret: string | null) => {
-	if (typeof val !== 'string') {
+	if (typeof val !== 'string')
 		throw new TypeError('Cookie value must be provided as a string.')
-	}
-	if (secret === null) {
-		throw new TypeError('Secret key must be provided.')
-	}
+
+	if (secret === null) throw new TypeError('Secret key must be provided.')
+
 	const encoder = new TextEncoder()
 	const secretKey = await crypto.subtle.importKey(
 		'raw',
@@ -471,16 +470,20 @@ export const signCookie = async (val: string, secret: string | null) => {
 		secretKey,
 		encoder.encode(val)
 	)
+
 	const hmacArray = Array.from(new Uint8Array(hmacBuffer))
 	const digest = btoa(String.fromCharCode(...hmacArray))
 	return `${val}.${digest.replace(/=+$/, '')}`
 }
 
 export const unsignCookie = async (input: string, secret: string | null) => {
-	if ('string' != typeof input)
+	if (typeof input !== 'string')
 		throw new TypeError('Signed cookie string must be provided.')
-	if (null == secret) throw new TypeError('Secret key must be provided.')
+
+	if (null === secret) throw new TypeError('Secret key must be provided.')
+
 	const tentativeValue = input.slice(0, input.lastIndexOf('.'))
 	const expectedInput = await signCookie(tentativeValue, secret)
+
 	return expectedInput === input ? tentativeValue : false
 }