+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.7.16 - 
 Improvement:
 - `t.Cookie` cookie option type
+- [#253](https://github.com/elysiajs/elysia/pull/253) platform agnostic cookie
+- Decorator like `state`, `decorate` and `derive`, doesn't apply to WebSocket `data`
 
 # 0.7.15 - 26 Sep 2023
 Change: